requirements.txt
Flask
google-generativeai
nltk
numpy


